✨ #15 카드 뽑기 페이지 -> 결과 페이지 이동 구현
prepareResult에 결과 로딩 대비 비동기 함수 추가(pageLoading.js를 따로 빼며), 카드 뽑기 페이지에서 결과 보기 버튼 클릭 시 해당 비동기 함수 실행되며 (swtichPage.js 확인) 결과 페이지로 이동
