+++ index.html
@@ -8,7 +8,7 @@
   </head>
   <body>
     <!-- 랜딩 페이지 -->
-    <section id="landing" class="active">
+    <section id="landing">
       <h1>삼타로</h1>
       <p>세 장의 카드로 해결하는 당신의 고민</p>
       <button id="start-btn">시작하기</button>
