+++ scripts/ui/switchPage.js
@@ -1,4 +1,5 @@
-import { prepareDrawingCards } from "../script.js";
+import { prepareDrawingCards } from "./prepareDrawingCards.js";
+import { prepareResult } from "./prepareResult.js";
 import { addDragScroll } from "./dragScroll.js";
 
 export function switchPage(fromId, toId) {
@@ -6,9 +7,7 @@ export function switchPage(fromId, toId) {
   const to = document.querySelector(`#${toId}`);
 
   from.classList.add("hidden");
-  from.classList.remove("active");
   to.classList.remove("hidden");
-  to.classList.add("active");
 
   if (toId === "draw") {
     // 페이지 전환 뒤 렌더링 시간... 이후 실행 (repaint 이후 실행 위함)
@@ -17,6 +16,12 @@ export function switchPage(fromId, toId) {
 
       const cardContainer = document.querySelector("#card-container");
       addDragScroll(cardContainer);
+    }, 0);
+  }
+
+  if (toId === "result") {
+    setTimeout(() => {
+      prepareResult();
     }, 0);
   }
 }
